move pages files into directories
