--- src/pages/admin/AdminDashboard.tsx
@@ -1,47 +1,0 @@
-import { Table, Tbody, Td, Th, Thead, Tr } from '@chakra-ui/react';
-import * as React from 'react';
-import { useQuery } from 'react-query';
-import apiClient from '../../api/apiClient';
-import User from '../../api/dtos/user.dto';
-import ErrorAlert from '../../components/ErrorAlert';
-import LoadingSpinner from '../../components/LoadingSpinner';
-
-const AdminDashboard: React.FC = () => {
-  const { isLoading, error, data } = useQuery<User[], Error>('adminList', () =>
-    apiClient.getAdmins(),
-  );
-
-  const getTableRows = (): React.ReactNode => (
-    <>
-      {data &&
-        data.map((admin) => (
-          <Tr>
-            <Td>Placeholder name</Td>
-            <Td>{admin.email}</Td>
-            <Td>Placeholder date</Td>
-          </Tr>
-        ))}
-    </>
-  );
-
-  return (
-    <>
-      {data && (
-        <Table variant="simple">
-          <Thead>
-            <Tr>
-              <Th>Admin</Th>
-              <Th>Email</Th>
-              <Th>Date Added</Th>
-            </Tr>
-          </Thead>
-          <Tbody>{getTableRows()}</Tbody>
-        </Table>
-      )}
-      {isLoading && <LoadingSpinner />}
-      {error && <ErrorAlert />}
-    </>
-  );
-};
-
-export default AdminDashboard;
