+++ src/pages/exampleFormPage/ExampleFormPage.test.tsx
@@ -1,6 +1,6 @@
-import { screen } from '@testing-library/react';
 import React from 'react';
-import { render } from '../test-utils';
+import { screen } from '@testing-library/react';
+import { render } from '../../test-utils';
 import ExampleFormPage from './ExampleFormPage';
 
 describe('ExampleFormPage', () => {
