+++ src/pages/researcherLandingPage/ResearcherLandingPage.test.tsx
@@ -2,17 +2,17 @@ import { screen } from '@testing-library/react';
 import React from 'react';
 import apiClient from '../../api/apiClient';
 import { render } from '../../test-utils';
-import AdminDashboard from './AdminDashboard';
+import ResearcherLandingPage from './ResearcherLandingPage';
 import User from '../../api/dtos/user.dto';
 
 jest.mock('../../api/apiClient');
 
 describe('Landing', () => {
-  test('renders admin dashboard list', async () => {
+  test('renders researcher landing page list of admins', async () => {
     apiClient.getAdmins = jest.fn(() =>
       Promise.resolve([{ id: 1, email: 'test@test.com', role: 'admin' }]),
     ) as () => Promise<User[]>;
-    render(<AdminDashboard />);
+    render(<ResearcherLandingPage />);
     const testEmail = await screen.findByText('test@test.com');
     expect(testEmail).toBeInTheDocument();
   });
